No separate /events file
