+++ Makefile
@@ -36,10 +36,8 @@ link:
 	@npm link
 
 test:
-	@echo "\nTesting the non-events file, hang on..."
+	@echo "\nTesting source files, hang on..."
 	@$(BIN)/mocha $(MOCHA_DEV)
-	@echo "\nTesting the events file, hang on..."
-	@TEST_EVENTS=1 $(BIN)/mocha $(MOCHA_DEV)
 
 test-watch:
 	@echo "\nStarting test watcher..."
