+++ lib/__tests__/lib.test.js
@@ -1,7 +1,5 @@
-const TEST_EVENTS = process.env.TEST_EVENTS == 1;
-
 // Use plain old commonjs require to ensure it works everywhere...
-const StyleSheet = require(TEST_EVENTS ? '../' : '../events');
+const StyleSheet = require('../');
 import assert from 'assert';
 import { parse } from 'css' ;
 import * as style from './fixtures';
@@ -10,6 +8,18 @@ const unique = arr => [...new Set(arr)];
 const values = obj => Object.keys(obj).map( key => obj[key] );
 
 describe('stilr', () => {
+  it('emits "update" every time create() is invoked / listens', (done) => {
+    const specialColorCode = '#a339d2';
+
+    StyleSheet.on('update', () => {
+      const newCSS = StyleSheet.render();
+      if (newCSS.indexOf(specialColorCode) !== -1) {
+        done();
+      }
+    });
+    StyleSheet.create({x: {color: specialColorCode}});
+  });
+
   it('exports the Map class used for making stylesheets', () => {
     assert.ok(StyleSheet.Map)
   });
