+++ lib/index.js
@@ -6,13 +6,17 @@ import {
   isEmpty,
 } from './utils';
 
+import EventEmitter from 'emmett';
+
+const emitter = new EventEmitter();
+
 let globalStylesheet = new Map();
 
 export default {
   create( styles, stylesheet = globalStylesheet ) {
     if ( !(stylesheet instanceof Map) ) throw new Error(`${ stylesheet } should be a Map`);
 
-    return Object.keys( styles ).reduce( ( acc, key ) => {
+    var rtn = Object.keys( styles ).reduce( ( acc, key ) => {
 
       let { style, pseudos, mediaQueries } = seperateStyles( styles[key] );
       const className = createClassName( sortObject( style ));
@@ -79,6 +83,14 @@ export default {
       acc[ key ] = className;
       return acc;
     }, {});
+
+    // Make sure this emitting step comes after the above stylesheet parsing
+    // step completes.
+    // Otherwise, when users access things like #render() inside the event
+    // callback, they may not get the newest stylesheet.
+    emitter.emit('update');
+
+    return rtn;
   },
 
   render(options = { pretty: false }, stylesheet = globalStylesheet ) {
@@ -112,6 +124,10 @@ export default {
   clear( stylesheet = globalStylesheet ) {
     stylesheet.clear();
     return !stylesheet.size;
+  },
+
+  on(...args) {
+    emitter.on(...args);
   },
 
   Map,
